Changed wait() to take a float param
- Allows more specific wait times

+++ BasicGameFramework/game.h
@@ -43,7 +43,7 @@ private:
 
 // utility funcs
 
-void wait(unsigned n);
+void wait(float n);
 void shuffleLoad();
 
 #endif
